+++ kipper/core/src/compiler/semantics/processor/type-checker.ts
@@ -27,7 +27,6 @@ import {
 	ArgumentTypeError,
 	ArithmeticOperationTypeError,
 	AssignmentTypeError,
-	FunctionReturnTypeError,
 	InvalidConversionTypeError,
 	InvalidRelationalComparisonTypeError,
 	InvalidUnaryExpressionTypeError,